Remove mock for mountutil tests
